refactoring hostname based action
