+++ cas-server-support-spnego/src/main/java/org/jasig/cas/support/spnego/web/flow/client/LdapSpnegoKnownClientSystemsFilterAction.java
@@ -113,7 +113,7 @@ public class LdapSpnegoKnownClientSystemsFilterAction extends BaseSpnegoKnownCli
     }
 
     @Override
-    protected boolean doSpnego() {
+    protected boolean shouldDoSpnego() {
         Connection connection = null;
         try {
             connection = createConnection();
